+++ mapchete/config.py
@@ -15,6 +15,7 @@ from collections import OrderedDict
 from copy import deepcopy
 from enum import Enum
 import fiona
+import fsspec
 import hashlib
 import importlib
 import inspect
@@ -1006,7 +1007,8 @@ def raw_conf(mapchete_file):
     if isinstance(mapchete_file, dict):
         return _map_to_new_config(mapchete_file)
     else:
-        return _map_to_new_config(yaml.safe_load(open(mapchete_file, "r").read()))
+        with fsspec.open(mapchete_file, "r") as src:
+            return _map_to_new_config(yaml.safe_load(src.read()))
 
 
 def raw_conf_process_pyramid(raw_conf, reset_pixelbuffer=False):
@@ -1281,6 +1283,20 @@ def _handle_deprecated_process_parameters(config_dict):
             out[k] = v
     return out
 
+
+def _config_to_dict(input_config):
+    if isinstance(input_config, dict):
+        if "config_dir" not in input_config:
+            raise MapcheteConfigError("config_dir parameter missing")
+        return OrderedDict(input_config, mapchete_file=None)
+    # from Mapchete file
+    elif os.path.splitext(input_config)[1] == ".mapchete":
+        with fsspec.open(input_config, "r") as config_file:
+            return OrderedDict(
+                yaml.safe_load(config_file.read()),
+                config_dir=os.path.dirname(os.path.realpath(input_config)),
+                mapchete_file=input_config,
+            )
 
 def _config_at_zoom(config, zooms):
     """Return parameter dictionary per zoom level."""